requirements.txt
# Django Core
django==5.0.1
python-decouple==3.8

# Database
# psycopg2-binary==2.9.9

# Forms & UI
# django-crispy-forms==2.2
# crispy-bootstrap5==0.7

# Image Processing
Pillow==10.2.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.0
python-docx==1.1.2
pytesseract==0.3.10

# AI/ML Integration
openai==1.14.3
requests==2.31.0

# Background Jobs & Queue
celery==5.3.6
# redis==5.0.3

# Cloud Storage (S3)
# boto3==1.34.48
# django-storages==1.14.4

# Virus Scanning
# clamd==1.0.2
# python-magic-bin==0.4.14

# Additional Utilities
numpy==1.26.4
# scipy==1.13.1
